First iteration of the full refactoring
This is the first stage of the refactoring of the fuel tank analysis.
This is the state till the USA. Some more testing is required and the
actual analysis are still to be performed. Moreover the code is to be
commented and cleaned. The most usefull is that this is to done when the
analysis have been done, as after this the conde should be with the
least possible amount of coupling.