Create different type of fuel flows
Before there was a single type of fuel flow. However, when the tank is
drained, the fuel being drained can not be defined, as this depends upon
the fuel in the tank. As such three different type of fuel flows are
created, which can be used as general mass flow, out, and in flowing
fuels.