Add get_phase method to hydrogen classes
The method has been added to enable to get the partition of the fuel
when giving the phase. This is to be used in draining analysis where a
fuel is to be drained, only knowing the phase of the fuel. This way
when there is a two phase fuel, the desired phase can be drained.