Refactor internal model
The internal model has been refacotred such that the only method that is
to be specified per model is the creation of the internal resistances.
This is beacuse it is assumed that the gas model is the same for the
different models. As such the code now should reduce the degree of
repertition.