Implememt tank mass computation
The computation of the tank mass has been performed. This used material
and the structural model factory.

In later stages it may be necessary to include the material in the
initialisation of the tank, this improves the computation of the tank
mass as well as the thermal capacity of the tank which is required for
the dynamic models.