Implement thermal capacity of fuel tank
The thermal capacity of the tank was not implemented. Now it has.

The convergence of the tank thermal capacity is to be implemented. This
means that an estimate for the tank mass is done, based on the operating
pressure; however, the actual operaring pressure changes depending on
the mission of the aircraft.