Clean mission definition and tank states
The mission has been defined in a seperate variable, to clean the code,
and reusability in other parts.

An indice has been create to avoid that the first tank state of a new
mission section is not added to the tank states, as this would mean
that there are double tank states in the object.