Refactor tank state derivatives computation
The code has been refacotred such that the state derivatives computation
has become a method of the tank state. Rather than using methdos to set
properties of the tank state such as the heat flux and the thermal
capacity of the tank state, these are passed in the same method.