Refactor mission and mission section analysis
The two classes have been refactored to enable the analysis of a mission
and mission sections. Before these were data classes, but this is not a
clean way to define the classes.

Now the classes have a method to perform the analys, which return the
tank states, which are to be used for the analysis.