Include mininum pressure in analysis
The minimum pressure has to be included to analyse tanks where the gas
phase is drained. For now only two phase tanks can be included, as these
use the fill level, the fill level is not used in single gas phase tanks
and thus is not accepted in the analysis this is to be included in later
staged.