Create module
The module is to be used to analyse the geometry of a tank. This creates
the contour plot for different combinations of radius and length. The
analyse_tank function can be used by providing different initial
conditions, insulation or fuel phase flow. This means that two-phase
tanks can be analysed as well as gas tanks.