Chnage InitialState to InitialConditions class
Before the InitialState of the thermodynamics module was used to define
the initial state of the tank. The goal is to eliminate the dependancy
of the analysis from these types of modules. Therefore an
InitialCondittions dataclass is created in the facades module, which is
to be used to define the initial conditions of the tank. This calss is
subsequently upacked by the facade.