Ensure null liquid mass when no fill
The liquid masses of the tank states are defined in a list comprehension
where it should be accounted for that when there is null fill, there is
no liquid hydrogen, thus no density can be requested. In the list
comprehension an if else statement has been added which checks the fill
and computes the liquid mass accordingly.