Change tank fill computation
The computation of the fill was soly based on the liquid mass change.
This caused problems in the switch fuel phase drain analysis, leading
to inconistencies in the change in total fuel. Now the fuel fill is
based on the total change in mass and then the densities of the new fuel
state.