Refactor efficiency computers
Before the efficiency computers were based on data classes, now they
have been changed to have a compute efficiency method wich is to be
called to compute the efficiency. This improves the integration with
the tank performance class, which with can be initialised with
different performance computers. Thus improving the initialisation of
the class objects.