Split gas and liquid analysis
Before there was a single file that did the gas and liquid tank
analysis. I have decided to split the files into a liquid analysis and a
gas analysis. This created a better overview of the files and the inputs
for the analysis. Before one needed to uncomment the desired sections,
now this is not needed anymore.