new dataclass for GenericTankDimensions
added a new dataclass to handle the Winnefeld definition of a tank. Inheriting from the TankDimension base class, we add parameters a (semi axis horizontal of shell x-sec) and b (semi axis longitudinal of ellipsoid) as input to describe a more general tank shape.

The tank factory is modified accordingly to create the right type of tank.