new tank sections and tank type added
The new WinnefeldTank inherits from the Tank base class in a manner similar to CylindricalTankSphericalCaps, but this time with the flexibility to have elliptic cylinder geometry for the shell and ellipsoidal geometry for the endcap.

The create_tank method is modified accordingly to accept this new tank type.

The new tank sections (EllipticCylinderBody and EllipsoidalEndCap) are modified to compute the the volume and surface areas of the new geometries. For now, the EllipticCylinder is the same as a regular cylinder (radius, c = a), but the EllipsoidalEndCap computes volume and area differently.

Volume computation for the ellipsoidal sections is straight forward, with a well known analytical volume formula applied. The surface area is more complex. Deriving an area from the known parameters would require the use of elliptic integrals, which I deemed overkill for the application at hand. Instead, I seed random points on the surface and triangulate the surface, taking the sum of the triangular areas as the wetted area. This operation is done with helper functions random_points_on_ellipsoid, filter_points_by_height and the scipy ConvexHull method. The resolution of the area depends on the selected discretization of the surface. For now this is hardcoded as 1000 randomly seeded points.