implemented first version of the optimization
In this framework, we are optimizing for the maximum gravimetric efficiency, where the tunable parameters are the geometric quantities: radius (AKA c) and the endcap length (b). The heuristic optimization used is of the Nelder-Mead kind. It remains to be determined if a gradient-based technique can be used here.

We plot the calculated gravimetric efficiencies as a function of psi (c/b) to study the sensitivity parameter.

TODO:
Next, I will implement a method for displaying the response surface where c and b are independent. I will also extend to a vacuum insulation type instead of rohacell.