refactor full cycle analysis: now using new {DensityAnalysisFacade} and new discharge, refuel, and dormancy MissionSections
there is something wrong with the refuel but I suspect it is due to the conditions I am forcing on it to reproduce the Stops paper. I believe they are using a cryopump which prevents the hydrogen from crossing the saturation line. I will put this on hold and focus on the discharge script.